Refactor ClusterClient into it's own file
Pulls all the logic around building/creating a ClusterClient into
a separate file.
