+++ src/lib.rs
@@ -420,6 +420,9 @@ pub mod geo;
 #[cfg_attr(docsrs, doc(cfg(feature = "cluster")))]
 pub mod cluster;
 
+#[cfg(feature = "cluster")]
+mod cluster_client;
+
 #[cfg(feature = "r2d2")]
 #[cfg_attr(docsrs, doc(cfg(feature = "r2d2")))]
 mod r2d2;
